+++ inst/ShinyStan/server_files/pages/diagnose/server/diagnostics.R
@@ -11,6 +11,12 @@ sp_nuts_check <- reactive({
     need(input$diagnostic_chain, message = "Loading...")
   )
 })
+lp_check <- reactive({
+  validate(
+    need(lp_name %in% dimnames(SAMPS_post_warmup)[[3]], 
+         message = "Plot not displayed\n(Draws for 'lp__' or 'log-posterior' not found)")
+  )
+})
 diagnostic_chain <- reactive({
   validate(need(input$diagnostic_chain, message = "Waiting for chain (0 for all)"))
   input$diagnostic_chain
@@ -36,7 +42,7 @@ selected_range <- debounce({
     "Divergence information" = "divergent"
   )
   input_nm <- paste0("dynamic_trace_diagnostic_", nm, "_out_date_window")
-  validate(need(input[[input_nm]], "Loading"))
+  validate(need(input[[input_nm]], "Updating selected range"))
   sel <- input[[input_nm]]
   high <- as.integer(strsplit(sel[[2]], "[-]")[[1]][1])
   low <- as.integer(if (is.nan(sel[[1]])) "1" else strsplit(sel[[1]], "[-]")[[1]][1])
@@ -65,6 +71,7 @@ dynamic_trace_diagnostic_stepsize <- reactive({
 })
 stepsize_vs_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   chain <- diagnostic_chain()
   sel <- selected_range()
   stepsize <- .stepsize_pw[if (!is.null(sel)) sel,-1L, drop = FALSE] # drop iterations column
@@ -97,6 +104,7 @@ stepsize_vs_accept_stat <- reactive({
 # sample (accept_stat, lp) ------------------------------------------------
 dynamic_trace_diagnostic_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   chain <- diagnostic_chain()
   samps <- SAMPS_post_warmup[, , lp_name]
   lab <- "Log Posterior"
@@ -130,6 +138,7 @@ dynamic_trace_diagnostic_accept_stat <- reactive({
 })
 lp_hist <- reactive({
   sp_nuts_check()
+  lp_check()
   chain <- diagnostic_chain()
   sel <- selected_range()
   lp <- SAMPS_post_warmup[if (!is.null(sel)) sel,, lp_name]
@@ -145,6 +154,7 @@ accept_stat_hist <- reactive({
 })
 accept_stat_vs_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   sel <- selected_range()
   metrop <- .accept_stat_pw[if (!is.null(sel)) sel, -1L, drop = FALSE] # drop iterations column
   lp <- SAMPS_post_warmup[if (!is.null(sel)) sel, , lp_name]
@@ -225,6 +235,7 @@ treedepth_ndivergent1_hist <- reactive({
 })
 treedepth_vs_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   chain <- diagnostic_chain()
   sel <- selected_range()
   treedepth <- .treedepth_pw[if (!is.null(sel)) sel,-1L, drop = FALSE] # drop iterations column
@@ -281,6 +292,7 @@ dynamic_trace_diagnostic_ndivergent <- reactive({
 })
 ndivergent_vs_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   chain <- diagnostic_chain()
   sel <- selected_range()
   ndivergent <- .ndivergent_pw[if (!is.null(sel)) sel,-1L, drop = FALSE] # drop iterations column
@@ -310,6 +322,32 @@ ndivergent_vs_accept_stat <- reactive({
 })
 
 
+
+# energy ------------------------------------------------------------------
+energy_hist <- reactive({
+  sp_nuts_check()
+  chain <- diagnostic_chain()
+  np <- bayesplot::nuts_params(SAMPLER_PARAMS_post_warmup, pars = "energy__")
+  if (chain != 0)
+    np <- subset(np, Chain == chain)
+  
+  schm <- unlist(bayesplot::color_scheme_get("brightblue"))
+  schm["light"] <- base_fill
+  schm["mid"] <- overlay_fill
+  schm["light_highlight"] <- vline_base_clr
+  schm["mid_highlight"] <- pt_outline_clr
+  bayesplot::color_scheme_set(unname(schm))
+  
+  bayesplot::mcmc_nuts_energy(np, merge_chains = isTRUE(chain != 0)) + 
+    ggplot2::facet_wrap(~ Chain, labeller = "label_both") + 
+    thm_no_yaxs + 
+    bayesplot::facet_bg(FALSE) + 
+    bayesplot::facet_text(size = rel(1)) + 
+    bayesplot::legend_move("right") + 
+    theme(legend.text.align = 0, legend.text = element_text(size = rel(1.5)))
+})
+
+
 # model parameter ---------------------------------------------------------
 dynamic_trace_diagnostic_parameter <- reactive({
   sp_nuts_check()
@@ -336,6 +374,7 @@ dynamic_trace_diagnostic_parameter <- reactive({
 })
 param_vs_lp <- reactive({
   sp_nuts_check()
+  lp_check()
   param <- diagnostic_param()
   chain <- diagnostic_chain()
   sel <- selected_range()
@@ -464,7 +503,8 @@ hmc_plots <- c("accept_stat_trace", "accept_stat_hist","accept_stat_vs_lp",
                "treedepth_vs_accept_stat", "ndivergent_vs_accept_stat", 
                "stepsize_vs_lp", "stepsize_vs_accept_stat", "stepsize_trace", 
                "param_vs_lp", "param_vs_accept_stat", "param_vs_stepsize", 
-               "param_vs_treedepth", "p_trace", "p_hist")
+               "param_vs_treedepth", "p_trace", "p_hist", 
+               "energy_hist")
 for (j in seq_along(trace_nms)) {
   local({
     fn <- paste0("dynamic_trace_diagnostic_", trace_nms[j])